backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pymupdf==1.24.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
pgvector==0.2.5
openai==1.30.1
httpx==0.27.0
beautifulsoup4==4.12.3
